# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/roles/my/community/{communityId}": {
      "get": {
        "operationId": "RolesController_getMyRolesForCommunity",
        "parameters": [
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRolesResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
    "/api/roles/my/channel/{channelId}": {
      "get": {
        "operationId": "RolesController_getMyRolesForChannel",
        "parameters": [
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRolesResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
...
    "/api/roles/user/{userId}/community/{communityId}": {
      "get": {
        "operationId": "RolesController_getUserRolesForCommunity",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRolesResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
    "/api/roles/user/{userId}/channel/{channelId}": {
      "get": {
        "operationId": "RolesController_getUserRolesForChannel",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRolesResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
    "/api/roles/user/{userId}/instance": {
      "get": {
        "operationId": "RolesController_getUserInstanceRoles",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRolesResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
    "/api/roles/community/{communityId}": {
      "get": {
        "operationId": "RolesController_getCommunityRoles",
        "parameters": [
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityRolesResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Roles"
        ]
      },
      "post": {
        "operationId": "RolesController_createCommunityRole",
        "parameters": [
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDto"
                }
              }
            }
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
    "/api/roles/community/{communityId}/{roleId}": {
      "put": {
        "operationId": "RolesController_updateRole",
        "parameters": [
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDto"
                }
              }
            }
          }
        },
        "tags": [
          "Roles"
        ]
      },
      "delete": {
        "operationId": "RolesController_deleteRole",
        "parameters": [
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
    "/api/roles/community/{communityId}/assign": {
      "post": {
        "operationId": "RolesController_assignRoleToUser",
        "parameters": [
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
    "/api/roles/community/{communityId}/users/{userId}/roles/{roleId}": {
      "delete": {
        "operationId": "RolesController_removeRoleFromUser",
        "parameters": [
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
    "/api/roles/community/{communityId}/{roleId}/users": {
      "get": {
        "operationId": "RolesController_getUsersForRole",
        "parameters": [
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleUserDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
...
    "/api/roles/instance": {
      "post": {
        "operationId": "RolesController_createInstanceRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDto"
                }
              }
            }
          }
        },
        "summary": "Create a new instance role",
        "tags": [
          "Roles"
        ]
      }
    },
    "/api/roles/instance/{roleId}": {
      "put": {
        "operationId": "RolesController_updateInstanceRole",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDto"
                }
              }
            }
          }
        },
        "summary": "Update an instance role",
        "tags": [
          "Roles"
        ]
      },
      "delete": {
        "operationId": "RolesController_deleteInstanceRole",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Delete an instance role",
        "tags": [
          "Roles"
        ]
      }
    },
    "/api/roles/instance/{roleId}/assign": {
      "post": {
        "operationId": "RolesController_assignInstanceRole",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Assign an instance role to a user",
        "tags": [
          "Roles"
        ]
      }
    },
    "/api/roles/instance/{roleId}/users/{userId}": {
      "delete": {
        "operationId": "RolesController_removeInstanceRole",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Remove an instance role from a user",
        "tags": [
          "Roles"
        ]
      }
    },
    "/api/roles/instance/{roleId}/users": {
      "get": {
        "operationId": "RolesController_getInstanceRoleUsers",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleUserDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get users assigned to an instance role",
        "tags": [
          "Roles"
        ]
      }
    },
...
    "/api/community/{id}": {
      "get": {
        "operationId": "CommunityController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Community"
        ]
      },
      "patch": {
        "operationId": "CommunityController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommunityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Community"
        ]
      },
      "delete": {
        "operationId": "CommunityController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Community"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kraken API",
    "description": "Kraken Chat Application API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RoleDto": {
        "type": "object",
        "properties": {
          "actions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "DELETE_MESSAGE",
                "DELETE_CHANNEL",
                "DELETE_COMMUNITY",
                "DELETE_INVITE",
                "DELETE_USER",
                "DELETE_ROLE",
                "DELETE_ALIAS_GROUP",
                "DELETE_ALIAS_GROUP_MEMBER",
                "DELETE_INSTANCE_INVITE",
                "DELETE_MEMBER",
                "DELETE_REACTION",
                "CREATE_MESSAGE",
                "CREATE_CHANNEL",
                "CREATE_COMMUNITY",
                "CREATE_INVITE",
                "CREATE_USER",
                "CREATE_ROLE",
                "CREATE_ALIAS_GROUP",
                "CREATE_ALIAS_GROUP_MEMBER",
                "CREATE_INSTANCE_INVITE",
                "CREATE_MEMBER",
                "CREATE_REACTION",
                "JOIN_CHANNEL",
                "READ_MESSAGE",
                "READ_CHANNEL",
                "READ_COMMUNITY",
                "READ_ALL_COMMUNITIES",
                "READ_USER",
                "READ_ROLE",
                "READ_ALIAS_GROUP",
                "READ_ALIAS_GROUP_MEMBER",
                "READ_INSTANCE_INVITE",
                "READ_MEMBER",
                "UPDATE_COMMUNITY",
                "UPDATE_CHANNEL",
                "UPDATE_MESSAGE",
                "UPDATE_USER",
                "UPDATE_ROLE",
                "UPDATE_ALIAS_GROUP",
                "UPDATE_ALIAS_GROUP_MEMBER",
                "UPDATE_INSTANCE_INVITE",
                "UPDATE_MEMBER",
                "CAPTURE_REPLAY",
                "READ_INSTANCE_SETTINGS",
                "UPDATE_INSTANCE_SETTINGS",
                "READ_INSTANCE_STATS",
                "MANAGE_USER_STORAGE",
                "BAN_USER",
                "KICK_USER",
                "TIMEOUT_USER",
                "UNBAN_USER",
                "PIN_MESSAGE",
                "UNPIN_MESSAGE",
                "DELETE_ANY_MESSAGE",
                "VIEW_BAN_LIST",
                "VIEW_MODERATION_LOGS"
              ]
            }
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "isDefault": {
            "type": "boolean"
          }
        },
        "required": [
          "actions",
          "id",
          "name",
          "createdAt",
          "isDefault"
        ]
      },
      "UserRolesResponseDto": {
        "type": "object",
        "properties": {
          "resourceType": {
            "type": "string",
            "enum": [
              "COMMUNITY",
              "CHANNEL",
              "INSTANCE"
            ],
            "nullable": true
          },
          "userId": {
            "type": "string"
          },
          "resourceId": {
            "type": "string",
            "nullable": true
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleDto"
            }
          }
        },
        "required": [
          "resourceType",
          "userId",
          "resourceId",
          "roles"
        ]
      },
      "CommunityRolesResponseDto": {
        "type": "object",
        "properties": {
          "communityId": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleDto"
            }
          }
        },
        "required": [
          "communityId",
          "roles"
        ]
      },
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "actions": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string",
              "enum": [
                "DELETE_MESSAGE",
                "DELETE_CHANNEL",
                "DELETE_COMMUNITY",
                "DELETE_INVITE",
                "DELETE_USER",
                "DELETE_ROLE",
                "DELETE_ALIAS_GROUP",
                "DELETE_ALIAS_GROUP_MEMBER",
                "DELETE_INSTANCE_INVITE",
                "DELETE_MEMBER",
                "DELETE_REACTION",
                "CREATE_MESSAGE",
                "CREATE_CHANNEL",
                "CREATE_COMMUNITY",
                "CREATE_INVITE",
                "CREATE_USER",
                "CREATE_ROLE",
                "CREATE_ALIAS_GROUP",
                "CREATE_ALIAS_GROUP_MEMBER",
                "CREATE_INSTANCE_INVITE",
                "CREATE_MEMBER",
                "CREATE_REACTION",
                "JOIN_CHANNEL",
                "READ_MESSAGE",
                "READ_CHANNEL",
                "READ_COMMUNITY",
                "READ_ALL_COMMUNITIES",
                "READ_USER",
                "READ_ROLE",
                "READ_ALIAS_GROUP",
                "READ_ALIAS_GROUP_MEMBER",
                "READ_INSTANCE_INVITE",
                "READ_MEMBER",
                "UPDATE_COMMUNITY",
                "UPDATE_CHANNEL",
                "UPDATE_MESSAGE",
                "UPDATE_USER",
                "UPDATE_ROLE",
                "UPDATE_ALIAS_GROUP",
                "UPDATE_ALIAS_GROUP_MEMBER",
                "UPDATE_INSTANCE_INVITE",
                "UPDATE_MEMBER",
                "CAPTURE_REPLAY",
                "READ_INSTANCE_SETTINGS",
                "UPDATE_INSTANCE_SETTINGS",
                "READ_INSTANCE_STATS",
                "MANAGE_USER_STORAGE",
                "BAN_USER",
                "KICK_USER",
                "TIMEOUT_USER",
                "UNBAN_USER",
                "PIN_MESSAGE",
                "UNPIN_MESSAGE",
                "DELETE_ANY_MESSAGE",
                "VIEW_BAN_LIST",
                "VIEW_MODERATION_LOGS"
              ]
            }
          },
          "name": {
            "type": "string",
            "maxLength": 50
          }
        },
        "required": [
          "actions",
          "name"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "actions": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string",
              "enum": [
                "DELETE_MESSAGE",
                "DELETE_CHANNEL",
                "DELETE_COMMUNITY",
                "DELETE_INVITE",
                "DELETE_USER",
                "DELETE_ROLE",
                "DELETE_ALIAS_GROUP",
                "DELETE_ALIAS_GROUP_MEMBER",
                "DELETE_INSTANCE_INVITE",
                "DELETE_MEMBER",
                "DELETE_REACTION",
                "CREATE_MESSAGE",
                "CREATE_CHANNEL",
                "CREATE_COMMUNITY",
                "CREATE_INVITE",
                "CREATE_USER",
                "CREATE_ROLE",
                "CREATE_ALIAS_GROUP",
                "CREATE_ALIAS_GROUP_MEMBER",
                "CREATE_INSTANCE_INVITE",
                "CREATE_MEMBER",
                "CREATE_REACTION",
                "JOIN_CHANNEL",
                "READ_MESSAGE",
                "READ_CHANNEL",
                "READ_COMMUNITY",
                "READ_ALL_COMMUNITIES",
                "READ_USER",
                "READ_ROLE",
                "READ_ALIAS_GROUP",
                "READ_ALIAS_GROUP_MEMBER",
                "READ_INSTANCE_INVITE",
                "READ_MEMBER",
                "UPDATE_COMMUNITY",
                "UPDATE_CHANNEL",
                "UPDATE_MESSAGE",
                "UPDATE_USER",
                "UPDATE_ROLE",
                "UPDATE_ALIAS_GROUP",
                "UPDATE_ALIAS_GROUP_MEMBER",
                "UPDATE_INSTANCE_INVITE",
                "UPDATE_MEMBER",
                "CAPTURE_REPLAY",
                "READ_INSTANCE_SETTINGS",
                "UPDATE_INSTANCE_SETTINGS",
                "READ_INSTANCE_STATS",
                "MANAGE_USER_STORAGE",
                "BAN_USER",
                "KICK_USER",
                "TIMEOUT_USER",
                "UNBAN_USER",
                "PIN_MESSAGE",
                "UNPIN_MESSAGE",
                "DELETE_ANY_MESSAGE",
                "VIEW_BAN_LIST",
                "VIEW_MODERATION_LOGS"
              ]
            }
          },
          "name": {
            "type": "string",
            "maxLength": 50
          }
        }
      },
      "AssignRoleDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "roleId": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "roleId"
        ]
      },
...
    }
  }
}